requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
gradio==4.0.0
Pillow==10.1.0
simple-lama-inpainting==0.0.1 